third_party/WebKit/LayoutTests/inspector/sources/debugger/debugger-pause-on-blocked-event-handler-expected.txt

CONSOLE ERROR: line 13: Refused to execute inline event handler because it violates the following Content Security Policy directive: "script-src 'self'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

Tests that pause on exception works for inline event handlers blocked by CSP.

Button
Script execution paused.
Call stack:
    0) testAction (pause-on-blocked-event-handler.js:4)
Paused on a script blocked due to Content Security Policy directive: "script-src 'self'".
Script execution resumed.
